+++ cli/src/commit_templater.rs
@@ -2789,6 +2789,7 @@ mod tests {
                 user_email: "test.user@example.com",
                 date_pattern_context: chrono::DateTime::UNIX_EPOCH.fixed_offset().into(),
                 default_ignored_remote: None,
+                use_glob_by_default: false,
                 extensions: &self.revset_extensions,
                 workspace: Some(RevsetWorkspaceContext {
                     path_converter: &self.path_converter,