+++ cli/tests/test_revset_output.rs
@@ -512,6 +512,44 @@ fn test_bad_symbol_or_argument_should_not_be_optimized_out() {
 }
 
 #[test]
+fn test_default_string_pattern() {
+    let test_env = TestEnvironment::default();
+    test_env.run_jj_in(".", ["git", "init", "repo"]).success();
+    let work_dir = test_env.work_dir("repo");
+
+    // substring match by default as of jj 0.35
+    let output = work_dir.run_jj(["log", "-rauthor('test.user')"]);
+    insta::assert_snapshot!(output.normalize_backslash(), @r"
+    @  qpvuntsm test.user@example.com 2001-02-03 08:05:07 e8849ae1
+    │  (empty) (no description set)
+    ~
+    [EOF]
+    ------- stderr -------
+    Warning: In revset expression
+     --> 1:8
+      |
+    1 | author('test.user')
+      |        ^---------^
+      |
+      = Default pattern syntax will be changed to `glob:` in a future release; use `substring:` prefix or set ui.revsets-use-glob-by-default=true to suppress this warning
+    [EOF]
+    ");
+
+    // with default flipped
+    let output = work_dir.run_jj([
+        "log",
+        "-rauthor('*test.user*')",
+        "--config=ui.revsets-use-glob-by-default=true",
+    ]);
+    insta::assert_snapshot!(output.normalize_backslash(), @r"
+    @  qpvuntsm test.user@example.com 2001-02-03 08:05:07 e8849ae1
+    │  (empty) (no description set)
+    ~
+    [EOF]
+    ");
+}
+
+#[test]
 fn test_alias() {
     let test_env = TestEnvironment::default();
     test_env.run_jj_in(".", ["git", "init", "repo"]).success();